requirements.txt
Django==2.1.2
django-bootstrap4==0.0.7
django-copyright==1.0.0
django-crispy-forms==1.7.2
django-isbn-field==0.5.2
django-request==1.5.4
django-request-vars==1.0.1
django-simple-history==2.4.0
django-suit==0.2.26
psycopg2-binary==2.7.5
python-dateutil==2.7.3
python-stdnum==1.9
pytz==2018.5
six==1.11.0
